Improve and reformat debian/copyright
Rewrite debian/copyright in the new proposed machine-parsable format.
Add copyright statements and a license for the schemas that are under
the W3C Software license.  Add a separate copyright holder for the
files that are owned by the Apache Software Foundation.  Reference the
Apache-2.0 license in /usr/share/common-licenses rather than including
the complete text.